bug fixes
found several bugs that were contributing to poor training performance. The biggest one: using i instead of j on lines 305 and 208 (adding the sum of gamma for observation j when hidden state i is equal to the observation, which makes no sense)

There are still optimization issues, the program can only use up to 30 hidden states due to memory over-consumption. I have an idea to fix this, but it might take some time.