pickling and dll
i plan to do the testing and analysis in python since there's more library support for that sort of thing there. As such, I plan on compiling a .dll with functions to test hmms that can be loaded from the file system and return results so that a library like sklearn can be used to analyze them.