scoring issues
After running some expirements using the training and testing framework developed so far, i found that the models trained and tested in the program universally underperform. I've tested the training code on well understood problems (i.e. the english language dataset) and found that it performed as expected. This leads me to believe that the problem lies more in the code for scoring/classifying sequences. One thing that I have not done enough research on is how to normalize scores with respect to sequence length.

I think that the log sequence probability metric i'm using right now is best suited for sequences of a fixed length, rather than variable length. I need to do more research in this area.